+++ automig/lib/diffing.py
@@ -11,8 +11,8 @@ def group_by_table(stmts):
   for stmt in stmts:
     if isinstance(stmt, (wrappers.CreateTable, wrappers.CreateIndex)):
       groups[stmt.table].append(stmt)
-    elif isinstance(stmt, wrappers.CreateEnum):
-      groups['enum', stmt.name].append(stmt)
+    elif isinstance(stmt, (wrappers.CreateEnum, wrappers.CreateExtension)):
+      groups[type(stmt), stmt.name].append(stmt)
     elif stmt is None:
       pass # where are these coming from? it happens in the test suite even
     else: